Move methods to food.cpp
