+++ CMakeLists.txt
@@ -60,6 +60,7 @@ set(CORE_SOURCES
     elonacore.cpp
     filesystem.cpp
     fish.cpp
+    food.cpp
     foobar_save.cpp
     get_card_info.cpp
     get_random_item_id.cpp
