+++ variables.hpp
@@ -26,6 +26,10 @@ ELONA_EXTERN(elona_vector1<int> ap);
 ELONA_EXTERN(int wishfilter);
 ELONA_EXTERN(int nooracle);
 
+// food.cpp
+ELONA_EXTERN(int enc);
+ELONA_EXTERN(int enc2);
+
 ELONA_EXTERN(elona_vector1<int> _randcolor);
 ELONA_EXTERN(elona_vector1<int> aniref);
 ELONA_EXTERN(elona_vector1<int> anix);
@@ -868,7 +872,6 @@ void cell_setchara(int = 0, int = 0, int = 0);
 void cell_swap(int = 0, int = 0, int = 0, int = 0);
 void chara_preparepic(int = 0, int = 0);
 void chara_vanquish(int = 0);
-void chara_vomit(int = 0);
 void check_kill(int = 0, int = 0);
 void check_quest();
 void clear_rprefmat();
@@ -1064,37 +1067,6 @@ void label_1579();
 void label_1580();
 void label_1581();
 void label_1583();
-void eat_rotten_food();
-void cook();
-void apply_general_eating_effect();
-void eating_effect_eat_iron();
-void eating_effect_insanity();
-void eating_effect_eat_horse();
-void eating_effect_eat_holy_one();
-void eating_effect_eat_at();
-void eating_effect_eat_guard();
-void eating_effect_fire();
-void eating_effect_insanity2();
-void eating_effect_eat_cute_one();
-void eating_effect_eat_lovely_one();
-void eating_effect_eat_poisonous_one();
-void eating_effect_regeneration();
-void eating_effect_eat_rotten_one();
-void eating_effect_strength();
-void eating_effect_magic();
-void eating_effect_insanity3();
-void eating_effect_calm();
-void eating_effect_insanity4();
-void eating_effect_chaos();
-void eating_effect_lightning();
-void eating_effect_eat_cat();
-void eating_effect_ether();
-void eating_effect_constitution();
-void eating_effect_magic2();
-void eating_effect_strength2();
-void eating_effect_will();
-void eating_effect_quick();
-void eating_effect_pregnant();
 void get_pregnant();
 void initialize_cell_object_data();
 void generate_random_nefia();
@@ -1263,8 +1235,8 @@ void spot_fishing();
 void spot_material();
 void spot_digging();
 void spot_mining_or_wall();
-void label_2160();
-void label_2161();
+void continuous_action_eating();
+void continuous_action_eating_finish();
 void label_2162();
 void label_2187();
 void label_2188();
@@ -1484,6 +1456,7 @@ void windowanime(int = 0, int = 0, int = 0, int = 0, int = 0, int = 0);
 void windowanimecorner(int = 0, int = 0, int = 0, int = 0, int = 0, int = 0);
 void cs_listbk();
 void cs_list(bool, const std::string&, int, int, int = 0, int = 0, int = 0);
+int access_character_info();
 
 template <typename T>
 void txt()
